#85 Checar se arma existe ou não antes de rodar o up/down

+++ database/migrations/2023_03_22_150809_delete_caution_weapon_table.php
@@ -11,11 +11,14 @@ return new class extends Migration {
      */
     public function up()
     {
-        $cautionWeapon = CautionWeapon::where('caution_id', 59)
-            ->where('weapon_description', 'PISTOLA TAURUS PT 640')
-            ->where('weapon_number', 'SDY03145')
-            ->first();
-        $cautionWeapon->delete();
+        if (
+            $cautionWeapon = CautionWeapon::where('caution_id', 59)
+                ->where('weapon_description', 'PISTOLA TAURUS PT 640')
+                ->where('weapon_number', 'SDY03145')
+                ->first()
+        ) {
+            $cautionWeapon->delete();
+        }
     }
 
     /**
@@ -25,16 +28,23 @@ return new class extends Migration {
      */
     public function down()
     {
-        $cautionWeapon = new CautionWeapon();
-        $cautionWeapon->caution_id = 59;
-        $cautionWeapon->entranced_at = '2023-02-14 11:45:00';
-        $cautionWeapon->exited_at = '2023-02-14 17:19:00';
-        $cautionWeapon->register_number = '';
-        $cautionWeapon->weapon_type_id = 1;
-        $cautionWeapon->weapon_description = 'PISTOLA TAURUS PT 640';
-        $cautionWeapon->weapon_number = 'SDY03145';
-        $cautionWeapon->cabinet_id = 1;
-        $cautionWeapon->shelf_id = 1;
-        $cautionWeapon->save();
+        if (
+            !($cautionWeapon = CautionWeapon::where('caution_id', 59)
+                ->where('weapon_description', 'PISTOLA TAURUS PT 640')
+                ->where('weapon_number', 'SDY03145')
+                ->first())
+        ) {
+            $cautionWeapon = new CautionWeapon();
+            $cautionWeapon->caution_id = 59;
+            $cautionWeapon->entranced_at = '2023-02-14 11:45:00';
+            $cautionWeapon->exited_at = '2023-02-14 17:19:00';
+            $cautionWeapon->register_number = '';
+            $cautionWeapon->weapon_type_id = 1;
+            $cautionWeapon->weapon_description = 'PISTOLA TAURUS PT 640';
+            $cautionWeapon->weapon_number = 'SDY03145';
+            $cautionWeapon->cabinet_id = 1;
+            $cautionWeapon->shelf_id = 1;
+            $cautionWeapon->save();
+        }
     }
 };
